% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{as_moment}
\alias{as_moment}
\title{Convert time class into a moment}
\usage{
as_moment(x, ...)
}
\arguments{
\item{x}{A time value to convert to a moment}
}
\description{
Convert time class into a moment
}
