% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_algebra.R
\name{calendar_algebra}
\alias{calendar_algebra}
\title{Calendrical algebra for converting between time units}
\usage{
calendar_algebra(x, y, ...)
}
\arguments{
\item{x}{The primary time unit}

\item{y}{The time unit to convert \code{x} into}

\item{at}{Optional time point for context-dependent ratios}
}
\value{
Numeric describing how many \code{x} time units fit into \code{y} at time \code{at}.
}
\description{
This S7 generic function defines the calendrical relationships between two
time units, and is the building block for defining calendars in mixtime. It
calculates how many \code{x} time units fit into one \code{y} unit. Some cyclical
granularities are context-dependent (such as the number of days in a month),
and so an optional time point can be provided with \code{at}.
}
\details{
The methods are dispatched based on the shortest path along defined methods.
This allows for defining only the direct relationships between adjacent
time units, and relying on graph traversal to find how to convert between
more distant units. For example the number of seconds in an hour can be
calculated from the number of seconds in a minute and then number of minutes
in an hour.

If a method is defined for converting between time units of different
calendar systems (e.g., Gregorian calendar days to Chinese calendar days),
then that method can be used to convert times at any granularity between the
two systems.
}
\examples{
# There are 12 months in a year
calendar_algebra(tu_year(1L), tu_month(1L))

# There are 7 days in a week
calendar_algebra(tu_week(1L), tu_day(1L))

# There are 3600 seconds in an hour
calendar_algebra(tu_hour(1L), tu_second(1L))

# There are 18 "2 months" in 3 years
calendar_algebra(tu_year(3L), tu_month(2L))

}
