% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar_data}
\alias{calendar_data}
\title{Extract the calendar data from an object}
\usage{
calendar_data(x)
}
\arguments{
\item{x}{An object containing a calendar}
}
\description{
\lifecycle{experimental}
}
\examples{
calendar_data(yearmonth(0:11))

}
