% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso8601.R
\name{tu_isoyear}
\alias{tu_isoyear}
\alias{tu_week}
\title{ISO 8601 time classes}
\usage{
tu_isoyear(.data = 1L)

tu_week(.data = 1L)
}
\arguments{
\item{.data}{The number of time units}
}
\value{
A time unit for the ISO 8601 "isoweek" calendar
}
\description{
ISO 8601 time classes
}
\examples{
yw <- linear_time(tu_week(1L), list(tu_year(1L)))
yw(Sys.Date())

}
