% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal0_time_civil.R
\name{calendar_time_civil}
\alias{calendar_time_civil}
\alias{cal_time_civil}
\title{Civil time unit classes}
\usage{
cal_time_civil
}
\value{
A time granule object for the civil time system.
}
\description{
Time unit constructors for the civil time system where the boundary of each
day is at midnight on the 24 hour clock. This calendar is intended to be
built on by other calendars (e.g. \verb{[cal_time_civil]} and \verb{[cal_isoweek]}) to
add common time components.  These units can be
used with \code{\link[=linear_time]{linear_time()}} to create custom time representations.
}
\details{
The following time units are available (\verb{cal_time_civil$}).
\itemize{
\item \code{day()}: Day unit
\item \code{hour()}: Hour unit
\item \code{minute()}: Minute unit
\item \code{second()}: Second unit
\item \code{millisecond()}: Millisecond unit
\item \code{microsecond()}: Microsecond unit
\item \code{nanosecond()}: Nanosecond unit
}
}
\examples{
# Create a custom time representation using civil time granules
hms <- new_cyclical_time_fn(
  chronon = second(1L),
  cycle = hour(1L)
)

}
\seealso{
\code{\link{cal_time_civil}}, \code{\link{cal_isoweek}}
}
