% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronon_convert.R
\name{chronon_convert}
\alias{chronon_convert}
\alias{chronon_convert.S7_methods}
\title{Convert between chronons}
\usage{
chronon_convert(x, ...)

chronon_convert.S7_methods(x, to, discrete = FALSE)
}
\arguments{
\item{x}{A linear time object (of class \code{mt_linear})}

\item{...}{Additional arguments for methods.
that \code{x} falls into is returned as an integer. If \code{FALSE}, a fractional
number of target chronons is returned (analagous to time using a continuous
time model).}

\item{to}{The target chronon to convert to (a time unit object)}

\item{discrete}{If \code{TRUE}, the number of target chronons since Unix epoch}
}
\value{
A numeric or integer representing the time since Unix epoch in terms of the target
chronon's precision.
}
\description{
Convert between chronons
}
\examples{

# Convert from months since epoch to years since epoch
chronon_convert(yearmonth(Sys.Date()), cal_gregorian$year(1L))

# Convert from days since epoch to months since epoch
chronon_convert(Sys.Date(), cal_gregorian$month(1L))
chronon_convert(Sys.Date(), cal_gregorian$month(1L), discrete = TRUE)

# Convert from seconds since epoch to hours since epoch
chronon_convert(Sys.time(), cal_gregorian$hour(1L))

}
