% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_epoch.R
\name{chronon_epoch}
\alias{chronon_epoch}
\title{Epoch offset for chronons}
\usage{
chronon_epoch(x, ...)
}
\arguments{
\item{x}{A chronon (time unit) object.}

\item{...}{Additional arguments for methods.}
}
\value{
A numeric value representing the epoch for the chronon.
}
\description{
Returns the epoch offset for a given chronon (time unit). The epoch defines
the starting point of the chronon's linear numbering, used when converting
from internal representations to common displays (e.g. applying an epoch of
1970-01-01).
}
\examples{
# The epoch for year linear time displays is 1970
chronon_epoch(cal_gregorian$year(1L))
}
