% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_format.R
\name{chronon_format_linear}
\alias{chronon_format_linear}
\alias{chronon_format_cyclical}
\title{Default formatting strings for chronons}
\usage{
chronon_format_linear(x, cal = time_calendar(x), ...)

chronon_format_cyclical(x, y, ...)
}
\arguments{
\item{x}{A time granule for the chronon.}

\item{cal}{The calendar of the chronon, used to disambiguate suitable format
strings for time units that are shared across calendars (e.g.
\code{cal_gregorian$day} and \code{cal_isoweek$day}).}

\item{...}{Additional arguments for methods.}

\item{y}{A time granule for the cycle}
}
\value{
A character string containing the default format template for the chronon.
}
\description{
Provides default linear time formatting strings for a given chronon (finest
time granule). The format strings use placeholders like \code{{lin(year(1L))}},
\verb{\{cyc(month(1L), year(1L)\}} and \verb{\{cyc(day(1L), month(1L)\}}, which are
evaluated in the context of the data's \code{\link[=time_calendar]{time_calendar()}}.
}
\examples{
chronon_format_linear(cal_gregorian$year(1L))
chronon_format_linear(cal_gregorian$month(1L))
chronon_format_linear(cal_gregorian$day(1L))
chronon_format_linear(cal_isoweek$day(1L))

chronon_format_cyclical(cal_gregorian$month(1L), cal_gregorian$year(1L))
chronon_format_cyclical(cal_gregorian$day(1L), cal_gregorian$month(1L))
chronon_format_cyclical(cal_isoweek$day(1L), cal_isoweek$week(1L))
chronon_format_cyclical(cal_isoweek$week(1L), cal_isoweek$year(1L))

}
