% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_format.R
\name{chronon_format_attr}
\alias{chronon_format_attr}
\title{Default formatting strings for chronon attributes}
\usage{
chronon_format_attr(x, ...)
}
\arguments{
\item{x}{A chronon (time unit) object.}

\item{...}{Additional arguments for methods.}
}
\value{
A character string containing the default format suffix for the chronon.
}
\description{
Provides suffixes for default formatting strings for a given chronon (time unit).
This provides useful information such as timezones or locations in the string.
}
