% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_granules.R
\name{chronon_granules}
\alias{chronon_granules}
\title{Default linear time granules for chronons}
\usage{
chronon_granules(x, ...)
}
\arguments{
\item{x}{A chronon (time unit) object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of granule values for the chronon. Returns an empty list
for the base \code{mt_unit} class.
}
\description{
Provides a default set of linear time granules for a given chronon (time unit).
Granules represent the subdivisions or breakpoints used when representing
time values at that chronon level.
}
\examples{
chronon_granules(cal_gregorian$year(1L))
chronon_granules(cal_gregorian$month(1L))
}
