% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{class_mixtime}
\alias{class_mixtime}
\title{Base S7 class for mixtime vector objects}
\usage{
class_mixtime(x = list(), i = seq_len(sum(lengths(x))))
}
\description{
\code{class_mixtime} is the base S7 class for all mixtime vector objects,
inheriting from \link[vecvec:class_vecvec]{vecvec::class_vecvec}. While not intended to be used
directly, this S7 class is suitable to use when defining S7 methods for
mixtime vectors. S3 methods can be defined using the \code{mixtime::mixtime}
class.
}
\seealso{
\code{\link[=mixtime]{mixtime()}} for creating mixtime vectors, and \code{\link[=new_mixtime]{new_mixtime()}} for
the low-level constructor function of this S7 class.
}
