% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_time.R
\name{continuous_time}
\alias{continuous_time}
\title{Continuous time representation}
\usage{
continuous_time(chronon, granules = list())
}
\arguments{
\item{chronon}{A time unit object representing the chronon (e.g., \code{tu_day(1)})}

\item{granules}{A list of time unit objects representing the granules (e.g., \code{list(tu_year(1), tu_month(1))})}
}
\value{
An function used to create continuous time points.
}
\description{
\code{continuous_time()} creates a continuous time representation using specified
granules and a chronon. Granules are larger time units that define the structure
of time (e.g., years, months), while the chronon is the smallest indivisible
time unit (e.g., days, hours).
}
\examples{

# A year-week time representation with weeks as the chronon
yw <- continuous_time(tu_week(1L), list(tu_year(1L)))
yw(Sys.Date())

# A year-month time representation with months as the chronon
ym <- continuous_time(tu_month(1L), list(tu_year(1L)))
ym(Sys.Date())

# A year-quarter-month time representation with months as the chronon
yqm <- continuous_time(tu_month(1L), list(tu_year(1L), tu_quarter(1L)))
yqm(1:100)
yqm(Sys.Date())

# A year-day time representation with days as the chronon
yd <- continuous_time(tu_day(1L), list(tu_year(1L)))
yd(Sys.Date())

}
