% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclical_time.R
\name{cyclical_gregorian}
\alias{cyclical_gregorian}
\alias{month_of_year}
\alias{day_of_year}
\alias{day_of_month}
\title{Gregorian cyclical time representations}
\usage{
month_of_year(.data, tz = NULL, discrete = TRUE)

day_of_year(.data, tz = NULL, discrete = TRUE)

day_of_month(.data, tz = NULL, discrete = TRUE)
}
\arguments{
\item{.data}{Another object to be coerced into the specified cyclical time.}

\item{tz}{Timezone, defaults to "UTC".}

\item{discrete}{If \code{TRUE}, the position within the cycle that \code{.data}
falls into is returned as an integer. If \code{FALSE}, a fractional
position is returned (analagous to time using a continuous time model).}
}
\description{
Cyclical time representations for the Gregorian calendar system. These functions
create time objects that repeat within a larger time cycle, useful for identifying
seasonal patterns or positions within a calendar period.
}
\details{
\itemize{
\item \code{month_of_year()}: Represents the month position within a year (1-12).
The chronon is one month, cycling within a year.
\item \code{day_of_year()}: Represents the day position within a year (1-365 or 1-366
for leap years). The chronon is one day, cycling within a year.
\item \code{day_of_month()}: Represents the day position within a month (1-28, 1-29,
1-30, or 1-31 depending on the month). The chronon is one day, cycling
within a month.
}

These cyclical representations are useful for analyzing seasonal patterns or
comparing time points at similar positions across different years.
}
\section{Custom Gregorian cyclical time representations}{

You can create custom cyclical time representations using \code{\link[=cyclical_time]{cyclical_time()}}
with any of the supported Gregorian time units (see \link{calendar_gregorian}).

For example, to create a representation for day of the month:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{day_of_month <- cyclical_time(
  chronon = cal_gregorian$day(1L),
  cycle = cal_gregorian$month(1L)
)
}\if{html}{\out{</div>}}
}

\examples{

month_of_year(Sys.Date())


day_of_year(Sys.Date())

}
\seealso{
\link{linear_gregorian} for linear Gregorian time representations,
\code{\link[=cyclical_time]{cyclical_time()}} for creating custom cyclical time representations
}
