% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclical_time.R
\name{cyclical_iso8601}
\alias{cyclical_iso8601}
\alias{day_of_week}
\alias{week_of_year}
\title{ISO 8601 day of week}
\usage{
day_of_week(.data, tz = NULL, discrete = TRUE)

week_of_year(.data, tz = NULL, discrete = TRUE)
}
\arguments{
\item{.data}{A vector to be coerced into day of week. This can be a date,
date-time, or numeric vector.}

\item{tz}{Timezone, defaults to "UTC".}

\item{discrete}{If \code{TRUE}, the number of chronons since Unix epoch that
\code{.data} falls into is returned as an integer. If \code{FALSE}, a fractional number
of chronons is returned (analagous to time using a continuous time model).}
}
\value{
A cyclical time object representing the day of the week.
}
\description{
A cyclical time representation for days within a week using the ISO 8601
standard where weeks start on Monday.
}
\examples{
day_of_week(Sys.Date())
day_of_week(as.Date("2025-12-15") + 0:6)

}
\seealso{
\code{\link[=yearweek]{yearweek()}} for ISO 8601 year-week representation,
\code{\link[=cyclical_time]{cyclical_time()}} for creating custom cyclical time representations
}
