% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_cyclical_labels.R
\name{cyclical_labels}
\alias{cyclical_labels}
\alias{cyclical_labels.S7_methods}
\title{Friendly labels for cyclical relationships}
\usage{
cyclical_labels(granule, cycle, ...)

cyclical_labels.S7_methods(granule, cycle, i)
}
\arguments{
\item{granule}{A time unit object representing the granule (e.g., \code{month(1L)})}

\item{cycle}{A time unit object representing the cycle (e.g., \code{year(1L)})}

\item{...}{Additional arguments for methods.}

\item{i}{Integer vector representing the position within the cycle.}
}
\value{
Character vector of labels for the time point within the cycle.
}
\description{
This S7 generic function provides the labels for cyclical relationships
between time units. These functions should return locale specific labels.
}
\examples{
# Labels for months in a year
with(cal_gregorian, cyclical_labels(month(1L), year(1L), 1:12))

# Labels for days in a week
with(cal_isoweek, cyclical_labels(day(1L), week(1L), 1:7))

# Labels for weeks in a year, defaulted from time_unit_abbr()
with(cal_isoweek, cyclical_labels(week(1L), year(1L), 1:52))

}
