% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclical_time.R
\name{cyclical_time}
\alias{cyclical_time}
\title{Cyclical time representation}
\usage{
cyclical_time(chronon, cycle, calendar = cal_gregorian)
}
\arguments{
\item{chronon}{A time unit object representing the chronon (e.g., \code{month(1L)})}

\item{cycle}{A time unit object representing the cycle (e.g., \code{year(1L)})}

\item{calendar}{A calendar used to find the time units (e.g., \code{cal_isoweek})}
}
\value{
An function used to create cyclical time points.
}
\description{
\code{cyclical_time()} creates a cyclical time representation using specified
cycle and chronon (and optionally display granules). The cycles is the
larger time unit that defines the time period over which the chronon loops
over (e.g. a week, \code{week(1L)}). The chronon is the smaller time unit that
iterates within each cycle (e.g. a day, \code{day(1L)}). Combined, these two
granules form a cyclical time point (e.g. day of the week).
}
\examples{

day_of_week <- cyclical_time(day(1L), week(1L), calendar = cal_isoweek)
day_of_week(Sys.Date())

month_of_year <- cyclical_time(month(1L), year(1L))
month_of_year(Sys.Date())

}
