% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{year}
\alias{year}
\alias{yearquarter}
\alias{yearmonth}
\title{Gregorian continuous time representations}
\usage{
year(.data, tz = "UTC")

yearquarter(.data, tz = "UTC")

yearmonth(.data, tz = "UTC")
}
\arguments{
\item{.data}{Another object to be coerced into the specified time.}

\item{tz}{Timezone, defaults to "UTC".}

\item{...}{Arguments for methods.}
}
\description{
\lifecycle{experimental}
}
\examples{

year(Sys.Date())


yearquarter(Sys.Date())


yearmonth(Sys.Date())

}
