% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{is_mixtime}
\alias{is_mixtime}
\title{Check if an object is a mixtime}
\usage{
is_mixtime(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A scalar logical: \code{TRUE} if \code{x} is a \code{mixtime} vector, \code{FALSE}
otherwise.
}
\description{
Tests whether \code{x} inherits from the \code{mixtime} class.
}
\examples{
is_mixtime(Sys.Date())
is_mixtime(mixtime(Sys.Date()))

}
\seealso{
\code{\link[=as_mixtime]{as_mixtime()}} to coerce objects to \code{mixtime},
\code{\link[=mixtime]{mixtime()}} to construct a \code{mixtime}.
}
