% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{is_moment}
\alias{is_moment}
\title{Check if the object is a moment}
\usage{
is_moment(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{moment} class.
}
\description{
Check if the object is a moment
}
\examples{
is_moment(Sys.Date())
is_moment(yearmonth(1))

}
