% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_time}
\alias{is_time}
\alias{time_is_linear}
\alias{time_is_cyclical}
\alias{time_is_duration}
\alias{is_time_linear}
\alias{is_time_cyclical}
\alias{is_time_duration}
\title{Check the time type of values}
\usage{
time_is_linear(x, ...)

time_is_cyclical(x, ...)

time_is_duration(x, ...)

is_time_linear(x, ...)

is_time_cyclical(x, ...)

is_time_duration(x, ...)
}
\arguments{
\item{x}{A time object (typically a \code{mixtime} vector).}

\item{...}{Additional arguments for methods.}
}
\value{
A logical vector the same length as \code{x}.
}
\description{
Test whether elements of a \code{mixtime} vector are linear, cyclical, or durations.
}
\details{
These helpers return a logical vector the same length as \code{x} identifying the
type of time represented by each element.
}
\examples{
t <- c(yearmonth(0), month_of_year(0), months(0L))
time_is_linear(t)
time_is_cyclical(t)
time_is_duration(t)

}
