% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{linear_gregorian}
\alias{linear_gregorian}
\alias{year}
\alias{yearquarter}
\alias{yearmonth}
\title{Gregorian continuous time representations}
\usage{
year(.data, tz = tz_name(.data), discrete = TRUE)

yearquarter(.data, tz = tz_name(.data), discrete = TRUE)

yearmonth(.data, tz = tz_name(.data), discrete = TRUE)
}
\arguments{
\item{.data}{Another object to be coerced into the specified time.}

\item{tz}{Timezone, defaults to "UTC".}

\item{discrete}{If \code{TRUE}, the number of chronons since Unix epoch that
\code{.data} falls into is returned as an integer. If \code{FALSE}, a fractional number
of chronons is returned (analagous to time using a continuous time model).}
}
\description{
Linear time representations for the Gregorian calendar system. These functions
create time objects measured in years, year-quarters, or year-months since the
Unix epoch (1970-01-01).
}
\details{
\itemize{
\item \code{year()}: Represents time in whole years since 1970. The chronon is one year.
\item \code{yearquarter()}: Represents time in quarters, grouped by year. The chronon
is one quarter, with years as the granule for display and grouping.
\item \code{yearmonth()}: Represents time in months, grouped by year. The chronon is
one month, with years as the granule for display and grouping.
}
}
\section{Custom Gregorian time representations}{

You can create custom time representations using \code{\link[=linear_time]{linear_time()}} with any of
the supported Gregorian time units (see \link{calendar_gregorian}).

For example, to create a time representation in hours since epoch with day granules:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dayhour <- linear_time(
  granules = list(tu_day(1L)),
  chronon = tu_hour(1L)
)
}\if{html}{\out{</div>}}
}

\examples{

year(Sys.Date())
year(Sys.Date(), discrete = FALSE)


yearquarter(Sys.Date())
yearquarter(Sys.Date(), discrete = FALSE)


yearmonth(Sys.Date())
yearmonth(Sys.Date(), discrete = FALSE)

}
