% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_linear_labels.R
\name{linear_labels}
\alias{linear_labels}
\alias{linear_labels.S7_methods}
\title{Friendly labels for linear relationships}
\usage{
linear_labels(granule, ...)

linear_labels.S7_methods(granule, i)
}
\arguments{
\item{granule}{A time unit object representing the granule (e.g., \code{year(1L)})}

\item{...}{Additional arguments for methods.}

\item{i}{Integer vector representing the position along the linear axis.}
}
\value{
Character vector of labels for the time point.
}
\description{
This S7 generic function provides the labels for linear (non-repeating)
positions of a time unit. These functions should return locale specific
labels.
}
\examples{
# Labels for years on a linear axis
with(cal_gregorian, linear_labels(year(1L), 2020:2025))

}
