% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime-package.R
\docType{package}
\name{mixtime-package}
\alias{mixtime-package}
\title{mixtime: Mixed Temporal Vectors and Operations}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Flexible time classes for time series analysis and forecasting with mixed temporal granularities. Supports linear and cyclical time representations in discrete and continuous forms, with timezone support, across multiple calendar systems including Gregorian and ISO week date calendars. Time points are stored numerically relative to a chronon; an atomic time granule defined by time units of a calendar. Calendrical arithmetic enables conversion between time granules (e.g. days to months) and calendar systems. Multi-unit arithmetic allows for temporal analysis with other granules of common calendars (e.g. fortnights are 2-week units). Time vectors of different granularities (e.g. monthly and quarterly) can be combined in a single vector, making 'mixtime' ideal for data that changes observation frequency over time or requires temporal reconciliation across scales. The package is extensible, allowing users to define custom calendars that build upon civil and astronomical time systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/mixtime/}
  \item \url{https://github.com/mitchelloharawild/mixtime}
  \item Report bugs at \url{https://github.com/mitchelloharawild/mixtime/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})
}

Other contributors:
\itemize{
  \item Cynthia Huang (\href{https://orcid.org/0000-0002-9218-987X}{ORCID}) [contributor]
  \item Rob Hyndman (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [contributor]
}

}
\keyword{internal}
