% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{mixtime_location}
\alias{mixtime_location}
\alias{loc_latitude}
\alias{loc_longitude}
\alias{loc_altitude}
\title{Extract locations from an object}
\usage{
loc_latitude(x, ...)

loc_longitude(x, ...)

loc_altitude(x, ...)
}
\arguments{
\item{x}{An object with location information.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric value representing the location (e.g., longitude, latitude, etc.).
}
\description{
Generic function to extract the location from objects that have location information.
}
\examples{
t <- linear_time(
  1:3, 
  cal_time_solar_sunset$day(1L, lat = -37.8136, lon = 144.9631)
)

loc_longitude(t)
loc_latitude(t)
loc_altitude(t)

}
