% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_in.R
\name{mixtime_nested}
\alias{mixtime_nested}
\title{Is a mixtime nested within another mixtime?}
\usage{
mixtime_nested(x, by)
}
\arguments{
\item{x}{A vector of mixtimes}

\item{by}{A vector of mixtimes for which \code{x} might be nested by}
}
\value{
A logical vector of same length as \code{x}.
}
\description{
\lifecycle{experimental}
}
\details{
This function helps identify if one temporal value is nested within the
granularity of another temporal value. For example, is 2020 Jan in 2020 Q1?
}
\examples{
mixtime_nested(yearmonth(0:5), by = yearquarter(0))
mixtime_nested(yearmonth(0:5), by = yearquarter(1))

}
