% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{mixtime_valid}
\alias{mixtime_valid}
\title{Check if times can be used within mixtime}
\usage{
mixtime_valid(x)
}
\arguments{
\item{x}{A vector of times.}
}
\value{
A logical vector.
}
\description{
Check if times can be used within mixtime
}
