% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment-package.R
\docType{package}
\name{moment-package}
\alias{moment}
\alias{moment-package}
\title{moment: Flexible time vectors}
\description{
Flexible time classes for representing mixed temporal granularities
    and custom calendar structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/moment/}
  \item \url{https://github.com/mitchelloharawild/moment}
  \item Report bugs at \url{https://github.com/mitchelloharawild/moment/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

}
\keyword{internal}
