% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{new_duration}
\alias{new_duration}
\title{Constructor for mixtime duration vectors}
\usage{
new_duration(x = integer(), chronon = NULL)
}
\arguments{
\item{x}{A numeric vector of time points, integers for discrete time or
doubles for continuous time.}

\item{chronon}{A time unit object representing the smallest indivisible time
unit (chronon) for the time vector (e.g. \code{cal_gregorian$day(1L)}).}
}
\value{
An object of class \code{mt_duration} representing the specified duration.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Constructs a new duration object (\code{mt_duration}) from a set of mixtime time
units. This functionality is in active development and is not ready for use.
}
\details{
This is a low-level constructor function that creates duration objects.
}
