% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{new_mixtime}
\alias{new_mixtime}
\title{Constructor for mixtime vectors}
\usage{
new_mixtime(x = mt_linear())
}
\arguments{
\item{x}{A mixtime time vector (created with \code{\link[=new_time]{new_time()}}) to wrap in a mixtime class.}
}
\value{
A \code{mixtime} object, which allows mixed-type time vectors to coexist in a single vector.
}
\description{
Creates a \code{mixtime} vector, which can contain time points of different
granularities (e.g. monthly and quarterly) in a single vector via \code{vecvec}.
}
