% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{new_moment}
\alias{new_moment}
\title{Create a new moment}
\usage{
new_moment(x = numeric(), calendar = new_calendar())
}
\arguments{
\item{x}{The iterable value of the moment (for example, a value of 1 may indicate the first day).}

\item{calendar}{The calendar structure over which the moment iterates. This can be created using \code{new_calendar()}.}
}
\description{
A moment is a vector which describes a point in time. It uses a calendar
definition to translate a vector of numbers into a point in time.
}
