% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_time}
\alias{round_time}
\alias{floor_time}
\alias{ceiling_time}
\alias{trunc_time}
\title{Round date/time objects to specified time units}
\usage{
round_time(x, unit, ...)
}
\arguments{
\item{x}{A date/time object to be rounded. Accepted types include Date,
POSIXct, POSIXlt and other objects that inherit from POSIXt. The returned
object will be of the same class as the input.}

\item{...}{Additional arguments passed to specific implementations.}

\item{units}{A time unit (or object coercible to a time unit, e.g. "day").}
}
\value{
An object of the same class as x with its time components adjusted
to the requested unit.
}
\description{
A family of helpers to round date/time objects to a specified time granule
such as second, minute, hour, or day. These functions preserve the input
time class, as rounded by the attributes of the \code{unit}.
}
\examples{
# Round POSIXct to the nearest minute (preserving tz)
t <- as.POSIXct("2020-01-01 12:34:56", tz = "UTC")
round_time(t, units = "minute")

# Floor to the nearest hour
floor_time(t, units = "hour")

# Ceiling a Date (treated as midnight-of-day rounding)
d <- as.Date("2020-01-01")
ceiling_time(d, units = "day")

}
\seealso{
\link[base:Round]{base::round}, \link[lubridate:round_date]{lubridate::floor_date}
}
