% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_unit.R
\name{time_unit<-}
\alias{time_unit<-}
\alias{set_time_units}
\title{Set the time units of a mixtime}
\usage{
time_unit(x) <- value

set_time_units(x, value)
}
\arguments{
\item{x}{A mixtime.}

\item{value}{The new time unit}
}
\description{
\lifecycle{experimental}
}
\details{
Modify a mixtime's granularity to use a different granularity. This is HIGHLY
experimental, and more of a proof of concept for how periods of time can be
combined by changing the granularity.
}
