% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_chronon.R
\name{time_chronon}
\alias{time_chronon}
\title{Obtain the chronon of a time object}
\usage{
time_chronon(x, ...)
}
\arguments{
\item{x}{A time object (e.g., \link[base:Dates]{base::Date}, \link[base:DateTimeClasses]{base::POSIXct}, \code{\link[=continuous_time]{continuous_time()}}, etc.)}
}
\value{
A time unit object representing the chronon (e.g., \code{tu_day(1L)})
}
\description{
This S7 generic function extracts the chronon (the smallest time unit) from a
time object, such as continuous time or cyclical time representations.
}
\examples{

# The chronon of a Date object is 1 day
time_chronon(Sys.Date())

# The chronon of a POSIXct object is 1 second
time_chronon(Sys.time())

# The chronon of a continuous time year and month is 1 month
time_chronon(continuous_time(tu_day(1L), list(tu_year(1L)))(Sys.Date()))
}
