% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian_time_units.R
\name{time_unit_ratio}
\alias{time_unit_ratio}
\title{Find the ratio between two time units}
\usage{
time_unit_ratio(x, y, ...)
}
\arguments{
\item{x}{First time unit}

\item{y}{Second time unit}

\item{at}{Optional time point for context-dependent ratios}
}
\value{
Numeric ratio of x to y
}
\description{
This S7 generic function calculates the ratio between two time units,
optionally at a specific time point.
}
\examples{
# Ratio of years to months
time_unit_ratio(tu_year(1), tu_month(1))


}
