% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{tz_abbreviation}
\alias{tz_abbreviation}
\title{Get timezone abbreviation}
\usage{
tz_abbreviation(time)
}
\arguments{
\item{time}{A POSIXct datetime object or something coercible to POSIXct.
The timezone is extracted from this object.}
}
\value{
A character vector of timezone abbreviations.
}
\description{
Returns the timezone abbreviation (e.g., "EST", "PDT") for a given datetime
in its specified timezone.
}
\examples{
tz_abbreviation(Sys.time())
tz_abbreviation(as.POSIXct("2024-01-15 12:00:00", tz = "America/New_York"))
}
