% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{tz_name}
\alias{tz_name}
\title{Extract timezone from an object}
\usage{
tz_name(x, ...)
}
\arguments{
\item{x}{An object with timezone information.}
}
\value{
A character string representing the timezone (e.g., "America/New_York", "UTC").
}
\description{
Generic function to extract the timezone from objects that have timezone information.
}
\examples{
tz_name(Sys.time())
tz_name(as.POSIXct("2024-06-15 12:00:00", tz = "America/New_York"))
}
