% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{tz_offset}
\alias{tz_offset}
\title{Get timezone offset}
\usage{
tz_offset(x, ...)
}
\arguments{
\item{x}{A time class coercible to POSIXt with an associated time zone.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{mixtime} duration vector of offsets from UTC in the same chronon
(e.g. seconds for POSIXt, days for dates, etc.)
}
\description{
Returns the UTC offset for a given datetime in its specified timezone.
}
\examples{
tz_offset(as.POSIXct(Sys.time(), tz = Sys.timezone()))
tz_offset(as.POSIXct("2024-06-15 12:00:00", tz = "America/New_York"))
}
