% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{year}
\alias{year}
\title{Represent years}
\usage{
year(x, ...)
}
\arguments{
\item{x}{Another object to be coerced into years}

\item{...}{Arguments for methods.}
}
\description{
Create or coerce using \code{year()}
}
\details{
\lifecycle{experimental}
}
