% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{yearmonth}
\alias{yearmonth}
\title{Represent yearmonths}
\usage{
yearmonth(x, ...)
}
\arguments{
\item{x}{Another object to be coerced into yearmonths}

\item{...}{Arguments for methods.}
}
\description{
Create or coerce using \code{yearmonth()}
}
\details{
\lifecycle{experimental}
}
