% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{yearquarter}
\alias{yearquarter}
\title{Represent yearquarters}
\usage{
yearquarter(x, ...)
}
\arguments{
\item{x}{Another object to be coerced into yearquarters}

\item{...}{Arguments for methods.}
}
\description{
Create or coerce using \code{yearquarter()}
}
\details{
\lifecycle{experimental}
}
