% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{yearweek}
\alias{yearweek}
\title{Represent yearweeks}
\usage{
yearweek(x, ...)
}
\arguments{
\item{x}{Another object to be coerced into yearweeks}

\item{...}{Arguments for methods.}
}
\description{
Create or coerce using \code{yearweek()}
}
\details{
\lifecycle{experimental}
}
