% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{as_mixtime}
\alias{as_mixtime}
\title{Convert time class into a mixtime}
\usage{
as_mixtime(x, ...)
}
\arguments{
\item{x}{A time value to convert to a mixtime}

\item{...}{Additional arguments for methods}
}
\description{
Convert time class into a mixtime
}
