% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{tu_year}
\alias{tu_year}
\alias{tu_quarter}
\alias{tu_month}
\alias{tu_day}
\alias{tu_hour}
\alias{tu_minute}
\alias{tu_second}
\alias{tu_millisecond}
\title{Gregorian time unit classes}
\usage{
tu_year(.data = 1L)

tu_quarter(.data = 1L)

tu_month(.data = 1L)

tu_day(.data = 1L)

tu_hour(.data = 1L)

tu_minute(.data = 1L)

tu_second(.data = 1L)

tu_millisecond(.data = 1L)
}
\arguments{
\item{.data}{The number of time units}
}
\value{
A time unit object for the Gregorian calendar system.
}
\description{
Time unit constructors for the Gregorian calendar system. These units can be
used with \code{\link[=linear_time]{linear_time()}} to create custom time representations.
}
\details{
The following Gregorian time units are available:
\itemize{
\item \code{tu_year()}: Year unit
\item \code{tu_quarter()}: Quarter (3-month period) unit
\item \code{tu_month()}: Month unit
\item \code{tu_day()}: Day unit
\item \code{tu_hour()}: Hour unit
\item \code{tu_minute()}: Minute unit
\item \code{tu_second()}: Second unit
\item \code{tu_millisecond()}: Millisecond unit
}

These units form a hierarchy where conversions between adjacent units follow
the Gregorian calendar rules. For units that don't have a fixed relationship
(e.g., months to days), the conversion requires a time context.
}
\examples{
# Create a custom time representation using Gregorian units
dayhour <- linear_time(
  granules = list(tu_day(1L)),
  chronon = tu_hour(1L)
)

}
\seealso{
\code{\link[=linear_time]{linear_time()}} for creating custom time representations,
\link{linear_gregorian} for pre-defined Gregorian time representations
}
