% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{tu_year}
\alias{tu_year}
\alias{tu_quarter}
\alias{tu_month}
\alias{tu_day}
\alias{tu_hour}
\alias{tu_minute}
\alias{tu_second}
\alias{tu_millisecond}
\title{Gregorian time unit classes}
\usage{
tu_year(x = integer(0))

tu_quarter(x = integer(0))

tu_month(x = integer(0))

tu_day(x = integer(0))

tu_hour(x = integer(0))

tu_minute(x = integer(0))

tu_second(x = integer(0))

tu_millisecond(x = integer(0))
}
\description{
Gregorian time unit classes
}
