% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{tu_year}
\alias{tu_year}
\alias{tu_quarter}
\alias{tu_month}
\alias{tu_day}
\alias{tu_hour}
\alias{tu_minute}
\alias{tu_second}
\alias{tu_millisecond}
\title{Gregorian time unit classes}
\usage{
tu_year(.data = 1L)

tu_quarter(.data = 1L)

tu_month(.data = 1L)

tu_day(.data = 1L)

tu_hour(.data = 1L)

tu_minute(.data = 1L)

tu_second(.data = 1L)

tu_millisecond(.data = 1L)
}
\arguments{
\item{.data}{The number of time units}
}
\value{
A time unit for the Gregorian calendar system.
}
\description{
Gregorian time unit classes
}
