% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{tu_year}
\alias{tu_year}
\alias{tu_quarter}
\alias{tu_month}
\alias{tu_day}
\alias{tu_hour}
\alias{tu_minute}
\alias{tu_second}
\alias{tu_millisecond}
\title{Gregorian time unit classes}
\usage{
tu_year(.data = integer(0))

tu_quarter(.data = integer(0))

tu_month(.data = integer(0))

tu_day(.data = integer(0))

tu_hour(.data = integer(0))

tu_minute(.data = integer(0))

tu_second(.data = integer(0))

tu_millisecond(.data = integer(0))
}
\description{
Gregorian time unit classes
}
