% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso8601.R
\name{calendar_iso8601}
\alias{calendar_iso8601}
\alias{tu_isoyear}
\alias{tu_week}
\title{ISO 8601 time unit classes}
\usage{
tu_isoyear(.data = 1L, tz = "UTC")

tu_week(.data = 1L, tz = "UTC")
}
\arguments{
\item{.data}{The number of time units}

\item{tz}{The timezone name for the unit (valid units can be found with \verb{[tzdb::tzdb_names()]})}
}
\value{
A time unit object for the ISO 8601 calendar system.
}
\description{
Time unit constructors for the ISO 8601 calendar system. These units can be
used with \code{\link[=linear_time]{linear_time()}} to create custom time representations.
}
\details{
The following ISO 8601 time units are available:
\itemize{
\item \code{tu_isoyear()}: ISO year unit (years start on the week containing the first Thursday)
\item \code{tu_week()}: Week unit (7-day periods)
}

ISO 8601 weeks always start on Monday and the first week of a year is the week
containing the first Thursday of that year. This means that some days in early
January may belong to the last week of the previous ISO year, and some days in
late December may belong to the first week of the next ISO year.
}
\seealso{
\code{\link[=linear_time]{linear_time()}} for creating custom time representations,
\code{\link[=yearweek]{yearweek()}} for a pre-defined ISO 8601 year-week representation
}
