% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso8601.R
\docType{data}
\name{calendar_iso8601}
\alias{calendar_iso8601}
\alias{cal_isoweek}
\title{ISO 8601 time unit classes}
\format{
An object of class \code{cal_isoweek} (inherits from \code{mt_calendar}) of length 7.
}
\usage{
cal_isoweek
}
\value{
A time unit object for the ISO 8601 calendar system.
}
\description{
Time unit constructors for the ISO 8601 calendar system. These units can be
used with \code{\link[=linear_time]{linear_time()}} to create custom time representations.
}
\details{
The following time units are available in the ISO week date calendar:
\itemize{
\item \code{year()}: ISO year unit (years start on the week containing the first Thursday)
\item \code{week()}: Week unit (7-day periods)
\item \code{day()}: Day unit
\item \code{hour()}: Hour unit
\item \code{minute()}: Minute unit
\item \code{second()}: Second unit
\item \code{millisecond()}: Millisecond unit
}

ISO 8601 weeks always start on Monday and the first week of a year is the week
containing the first Thursday of that year. This means that some days in early
January may belong to the last week of the previous ISO year, and some days in
late December may belong to the first week of the next ISO year.
}
\seealso{
\code{\link[=linear_time]{linear_time()}} for creating custom time representations,
\code{\link[=yearweek]{yearweek()}} for a pre-defined ISO 8601 year-week representation
}
\keyword{datasets}
