% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal0_time_civil.R
\docType{data}
\name{calendar_time_civil}
\alias{calendar_time_civil}
\alias{cal_time_civil_midnight}
\title{Civil time unit classes}
\format{
An object of class \code{cal_time_civil_midnight} (inherits from \code{mt_calendar}) of length 5.
}
\usage{
cal_time_civil_midnight
}
\value{
A time unit object for the civil time system.
}
\description{
Time unit constructors for the civil time system where the boundary of each
day is at midnight on the 24 hour clock. This calendar is intended to be
built on by other calendars (e.g. \verb{[cal_time_civil_midnight]} and \verb{[cal_isoweek]}) to
add common time components.  These units can be
used with \code{\link[=linear_time]{linear_time()}} to create custom time representations.
}
\details{
The following time units are available (\verb{cal_time_civil_midnight$}).
\itemize{
\item \code{day()}: Day unit
\item \code{hour()}: Hour unit
\item \code{minute()}: Minute unit
\item \code{second()}: Second unit
\item \code{millisecond()}: Millisecond unit
}
}
\examples{
# Create a custom time representation using Gregorian units
hms <- new_cyclical_time_fn(
  chronon = second(1L),
  chronon = hour(1L)
)

}
\seealso{
\code{\link{cal_time_civil_midnight}}, \code{\link{cal_isoweek}}
}
\keyword{datasets}
