% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal0_time_lunar.R
\docType{data}
\name{calendar_time_lunar}
\alias{calendar_time_lunar}
\alias{cal_time_lunar}
\title{Lunar time unit classes}
\format{
A location-based calendar containing lunar time units.
}
\usage{
cal_time_lunar
}
\description{
Time unit constructors for the lunar time system where the boundary of each
day is at sunrise, sunset, or noon. This calendar is intended to be
built on by other calendars to add common time components.
}
\details{
The following time units are available in the lunar calendar systems.
\itemize{
\item \code{month()}: Synodic month unit
\item \code{phase()}: Synodic phase unit
}
}
\examples{
# Find the time of a new moon in the Gregorian calendar
t <- linear_time(Sys.Date(), cal_time_lunar$month(1L, lat = -37.8136, lon = 144.9631))
datetime(t, tz = "Australia/Melbourne")


}
\seealso{
\code{\link{cal_time_civil}}
}
\keyword{datasets}
