% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal0_time_lunar.R, R/cal0_time_solar.R
\docType{data}
\name{calendar_time_solar}
\alias{calendar_time_solar}
\alias{cal_time_lunar}
\alias{cal_time_solar_sunset}
\alias{cal_time_solar_sunrise}
\alias{cal_time_solar_noon}
\alias{cal_time_solar_midnight}
\alias{cal_time_solar_dawn}
\alias{cal_time_solar_dusk}
\title{Lunar time unit classes}
\format{
An object of class \code{mt_calendar} of length 2.

An object of class \code{mt_calendar} of length 1.

An object of class \code{mt_calendar} of length 1.

An object of class \code{mt_calendar} of length 1.

An object of class \code{mt_calendar} of length 1.

An object of class \code{mt_calendar} of length 1.

An object of class \code{mt_calendar} of length 1.
}
\usage{
cal_time_lunar

cal_time_solar_sunset

cal_time_solar_sunrise

cal_time_solar_noon

cal_time_solar_midnight

cal_time_solar_dawn

cal_time_solar_dusk
}
\value{
A time unit object for the Gregorian calendar system.

A time unit object for the Gregorian calendar system.
}
\description{
Time unit constructors for the lunar time system where the boundary of each
day is at sunrise, sunset, or noon. This calendar is intended to be
built on by other calendars to add common time components.

Time unit constructors for the solar time system where the boundary of each
day is at sunrise, sunset, or noon. This calendar is intended to be
built on by other calendars to add common time components.
}
\details{
The following time units are available in the lunar calendar systems.
\itemize{
\item \code{month()}: Synodic month unit
\item \code{phase()}: Synodic phase unit
}

The following time units are available in the solar calendar systems.
\itemize{
\item \code{day()}: Day unit
}
}
\examples{
# Find the time of a new moon in the Gregorian calendar
t <- linear_time(Sys.Date(), cal_time_lunar$phase(0, lat = -37.8136, lon = 144.9631))
datetime(t, tz = "Australia/Melbourne")


# Find the time of sunset in the Gregorian calendar
t <- linear_time(Sys.Date(), cal_time_solar_sunset$day(1L, lat = -37.8136, lon = 144.9631))
datetime(t, tz = "Australia/Melbourne")

}
\seealso{
\code{\link{cal_time_civil_midnight}}

\code{\link{cal_time_civil_midnight}}
}
\keyword{datasets}
