% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronon_common.R
\name{chronon_common}
\alias{chronon_common}
\title{Find a common chronon from a set of chronons}
\usage{
chronon_common(..., .ptype = NULL)
}
\arguments{
\item{...}{A set of chronons to find a common chronon for.}

\item{.ptype}{If NULL, the default, the output returns the common chronon
across all elements of \code{...}. Alternatively, a prototype chronon can be
supplied to \code{.ptype} to demand a specific chronon is used. If the supplied
\code{.ptype} cannot represent all input chronons without loss of information,
an error is raised.}
}
\value{
A time unit object representing the common chronon.
}
\description{
This utility function takes a set of chronons and identifies a common chronon
of the finest granularity that can represent all input chronons without loss
of information. This is useful for operations that require a shared time
unit, such as combining or comparing different time measured at different
precisions.
}
\examples{

# Find common chronon between days and weeks
chronon_common(tu_day(1L), tu_week(1L))

# Find common chronon between weeks, and months
chronon_common(tu_week(1L), tu_month(1L))

# # Find common chronon between hours, months, and years
chronon_common(tu_hour(1L), tu_month(1L), tu_year(1L))

# Find common chronon between months, quarters, and years
chronon_common(tu_month(1L), tu_quarter(1L), tu_year(1L))

}
