% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronon_convert.R
\name{chronon_convert}
\alias{chronon_convert}
\title{Convert between chronons}
\usage{
chronon_convert(x, ...)
}
\arguments{
\item{x}{A linear time object (of class \code{mt_linear})}

\item{to}{The target chronon to convert to (a time unit object)}

\item{discrete}{If \code{TRUE}, the number of target chronons since Unix epoch
that \code{x} falls into is returned as an integer. If \code{FALSE}, a fractional
number of target chronons is returned (analagous to time using a continuous
time model).}
}
\value{
A numeric or integer representing the time since Unix epoch in terms of the target
chronon's precision.
}
\description{
Convert between chronons
}
\examples{

# Convert from months since epoch to years since epoch
chronon_convert(yearmonth(Sys.Date()), tu_year(1L))

# Convert from days since epoch to months since epoch
chronon_convert(Sys.Date(), tu_month(1L))
chronon_convert(Sys.Date(), tu_month(1L), discrete = TRUE)

# Convert from seconds since epoch to hours since epoch
chronon_convert(Sys.time(), tu_hour(1L))

}
