% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_divmod.R
\name{chronon_divmod}
\alias{chronon_divmod}
\alias{chronon_divmod,mixtime::mt_unit,mixtime::mt_unit-method}
\title{Convert between chronons of different time units}
\usage{
chronon_divmod(from, to, ...)

## S7 method for classes <mixtime::mt_unit>, <mixtime::mt_unit>
chronon_divmod(from, to, x)
}
\arguments{
\item{from}{The time unit that \code{x} is measured in (e.g., \code{day(1L)}).}

\item{to}{The time unit to convert \code{x} into (e.g., \code{week(1L)}).}

\item{...}{Additional arguments for methods.}

\item{x}{An integer vector of chronons measured in the \code{from} time unit.}
}
\value{
An list of two elements:
\itemize{
\item \code{div}: integer vector of chronons measured in the \code{to} time unit.
\item \code{mod}: integer vector of the remainder (in \code{from} time unit) after
converting to the \code{to} time unit.
}
}
\description{
This function converts between chronons measured in different time units. It
is used internally for converting between different continuous time types,
and is particularly useful for efficiently converting between irregular time
units. The default method uses \code{chronon_cardinality()} to cast between time
units, which is efficient for regular time units.
}
\examples{
# Convert day 16 after epoch (1970-01-01) into weeks since epoch (and remainder days)
with(cal_isoweek, chronon_divmod(day(1L), week(1L), 16L))

# Convert week 4 after epoch (1970-W1) into days since epoch
with(cal_isoweek, chronon_divmod(week(1L), day(1L), 4L))

}
