% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_format.R
\name{chronon_format_linear}
\alias{chronon_format_linear}
\alias{chronon_format_cyclical}
\title{Default formatting strings for chronons}
\usage{
chronon_format_linear(x, ...)

chronon_format_cyclical(x, y, ...)
}
\arguments{
\item{x}{A chronon (time unit) object.}

\item{...}{Additional arguments for methods.}

\item{y}{A chronon (time unit) for the cycle size.}
}
\value{
A character string containing the default format template for the chronon.
}
\description{
Provides default linear time formatting strings for a given chronon (time unit).
The format strings use placeholders like \code{{year}}, \code{{month}}, \code{{day}}, etc.,
that can be interpolated with actual values.
}
\examples{
chronon_format_linear(cal_gregorian$year(1L))
chronon_format_linear(cal_gregorian$month(1L))
chronon_format_linear(cal_gregorian$day(1L))

}
