% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_chronon_format.R
\name{chronon_format}
\alias{chronon_format}
\title{Default formatting strings for chronons}
\usage{
chronon_format(x, ...)
}
\arguments{
\item{x}{A chronon (time unit) object.}
}
\value{
A character string containing the default format template for the chronon.
}
\description{
Provides default linear time formatting strings for a given chronon (time unit).
The format strings use placeholders like \code{{year}}, \code{{month}}, \code{{day}}, etc.,
that can be interpolated with actual values.
}
\examples{
chronon_format(cal_gregorian$year)
chronon_format(cal_gregorian$month)
chronon_format(cal_gregorian$day)
}
