% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_time.R
\name{continuous_time}
\alias{continuous_time}
\title{Continuous time representation}
\usage{
continuous_time(granules = list(), chronon)
}
\arguments{
\item{granules}{A list of time unit objects representing the granules (e.g., \code{list(tu_year(1), tu_month(1))})}

\item{chronon}{A time unit object representing the chronon (e.g., \code{tu_day(1)})}
}
\value{
An S7 class used to create continuous time points.
}
\description{
\code{continuous_time()} creates a continuous time representation using specified
granules and a chronon. Granules are larger time units that define the structure
of time (e.g., years, months), while the chronon is the smallest indivisible
time unit (e.g., days, hours).
}
\examples{

# A yearmonth time representation with months as the chronon
yw <- continuous_time(list(tu_year(1L)), tu_week(1L))
yw(Sys.Date())

yqm <- continuous_time(list(tu_year(1L), tu_quarter(1L)), tu_month(1L))
yqm(1234)

yd <- continuous_time(list(tu_year(1L)), tu_day(1L))
yd(Sys.Date())

}
