% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclical_labels.R
\name{cyclical_labels}
\alias{cyclical_labels}
\title{Friendly labels for cyclical relationships}
\usage{
cyclical_labels(granule, cycle, ...)
}
\arguments{
\item{granule}{A time unit object representing the granule (e.g., \code{tu_month(1L)})}

\item{cycle}{A time unit object representing the cycle (e.g., \code{tu_year(1L)})}
}
\value{
Character vector of labels for the time point within the cycle.
}
\description{
This S7 generic function provides the labels for cyclical relationships
between time units. These functions should return locale specific labels.
}
\examples{
# Labels for months in a year
cyclical_labels(tu_month(1L), tu_year(1L), 1:12)

# Labels for days in a week
cyclical_labels(tu_day(1L), tu_week(1L), 1:7)

# Labels for weeks in a year, defaulted from time_unit_abbr()
cyclical_labels(tu_week(1L), tu_year(1L), 1:52)

}
