% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{new_duration}
\alias{new_duration}
\title{Create a temporal durations from time units}
\usage{
new_duration(...)
}
\arguments{
\item{...}{A set of time unit objects (e.g., \code{tu_year(1)}, \code{tu_month(2)}, etc.)}
}
\value{
An object of class \code{mt_duration} representing the specified duration.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Constructs a new duration object (\code{mt_duration}) from a set of mixtime time
units. This functionality is in active development and is not ready for use.
}
\details{
This is a low-level constructor function that creates duration objects.
}
