% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gregorian.R
\name{year}
\alias{year}
\alias{yearquarter}
\alias{yearmonth}
\title{Gregorian continuous time representations}
\usage{
year(.data, tz = NULL, discrete = TRUE)

yearquarter(.data, tz = NULL, discrete = TRUE)

yearmonth(.data, tz = NULL, discrete = TRUE)
}
\arguments{
\item{.data}{Another object to be coerced into the specified time.}

\item{tz}{Timezone, defaults to "UTC".}

\item{discrete}{If \code{TRUE}, the number of chronons since Unix epoch that
\code{.data} falls into is returned as an integer. If \code{FALSE}, a fractional number
of chronons is returned (analagous to time using a continuous time model).}

\item{...}{Arguments for methods.}
}
\description{
\lifecycle{experimental}
}
\examples{

year(Sys.Date())


yearquarter(0:7)


yearmonth(Sys.Date())

}
