% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{is_mixtime}
\alias{is_mixtime}
\title{Check if the object is a mixtime}
\usage{
is_mixtime(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{mixtime} class.
}
\description{
Check if the object is a mixtime
}
\examples{
is_mixtime(Sys.Date())
is_mixtime(yearmonth(1))

}
