% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_time.R
\name{linear_time}
\alias{linear_time}
\title{Linear time representation}
\usage{
linear_time(chronon, granules = list(), calendar = cal_gregorian)
}
\arguments{
\item{chronon}{A time unit object representing the chronon (e.g., \code{day(1)})}

\item{granules}{A list of time unit objects representing the granules (e.g., \code{list(year(1), month(1))})}

\item{calendar}{A calendar used to find the time units (e.g., \code{cal_isoweek})}
}
\value{
An function used to create continuous time points.
}
\description{
\code{linear_time()} creates a linear time representation using specified
granules and a chronon. Granules are larger time units that define the structure
of time (e.g., years, months), while the chronon is the smallest indivisible
time unit (e.g., days, hours).
}
\examples{

# A year-month time representation with months as the chronon
ym <- linear_time(month(1L), list(year(1L)))
ym(Sys.Date())

# A year-quarter-month time representation with months as the chronon
yqm <- linear_time(month(1L), list(year(1L), quarter(1L)))
yqm(1:100)
yqm(Sys.Date())

# A year-day time representation with days as the chronon
yd <- linear_time(day(1L), list(year(1L)))
yd(Sys.Date())

ymd_h <- linear_time(hour(1L), list(year(1L), month(1L), day(1L)))
ymd_h(Sys.time())

}
