% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime-package.R
\docType{package}
\name{mixtime-package}
\alias{mixtime}
\alias{mixtime-package}
\title{mixtime: Mixed Temporal Vectors and Operations}
\description{
Flexible time classes for representing mixed temporal granularities and bespoke calendar structures. Supports both linear and circular time representations in discrete and continuous forms across a variety of calendar systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/mixtime/}
  \item \url{https://github.com/mitchelloharawild/mixtime}
  \item Report bugs at \url{https://github.com/mitchelloharawild/mixtime/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
