% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime-package.R
\docType{package}
\name{mixtime-package}
\alias{mixtime}
\alias{mixtime-package}
\title{mixtime: Mixed time vectors}
\description{
Flexible time classes for representing mixed temporal granularities and custom calendar structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/mixtime/}
  \item \url{https://github.com/mitchelloharawild/mixtime}
  \item Report bugs at \url{https://github.com/mitchelloharawild/mixtime/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

}
\keyword{internal}
