% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{mt_unit}
\alias{mt_unit}
\alias{mt_tz_unit}
\title{Base S7 class for creating new time units}
\usage{
mt_unit(.data = 1L)

mt_tz_unit(.data = 1L, tz = "UTC")
}
\arguments{
\item{.data}{The number of time units}

\item{tz}{The timezone name for the unit (valid units can be found with \verb{[tzdb::tzdb_names()]})}
}
\value{
A time unit object of class \code{mt_unit}
}
\description{
This class is the primative class for time units, and should
be extended from when creating new time units. A new class
is typically created with S7 using:
\code{tu_day <- S7::new_class("tu_day", parent = mt_unit)}
}
