% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{data}
\name{mt_unit_s3}
\alias{mt_unit_s3}
\alias{mt_unit}
\title{Base S7 class for time units}
\format{
An object of class \code{S7_S3_class} of length 3.
}
\usage{
mt_unit_s3

mt_unit(.data = 1L)
}
\description{
Base S7 class for time units
}
\keyword{datasets}
