% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{new_calendar}
\alias{new_calendar}
\title{Create a new calendar}
\usage{
new_calendar(granularity = list_of_time_units(), origin = logical())
}
\arguments{
\item{granularity}{An origin-less identifier for the time increment of each value.}

\item{origin}{Logical identifier for if the origin is meaningful.}
}
\value{
A calendar data frame.
}
\description{
\lifecycle{experimental}
}
\details{
A calendar describes how time changes over some vector. It is responsible
for converting a vector of numbers into a point in time.
}
