% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{new_calendar}
\alias{new_calendar}
\title{Create a new calendar}
\usage{
new_calendar(origin, granularity)
}
\arguments{
\item{origin}{The initial time point for the calendar}

\item{granularity}{An origin-less identifier for the time increment of each value.}
}
\value{
A calendar data frame.
}
\description{
\lifecycle{experimental}
}
\details{
A calendar describes how time changes over some vector. It is responsible
for converting a vector of numbers into a point in time.
}
