% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{new_mixtime}
\alias{new_mixtime}
\title{Create a new mixtime}
\usage{
new_mixtime(x = list())
}
\arguments{
\item{x}{The list of time values to become a mixtime.}
}
\description{
A mixtime is a vector which describes a point in time. It uses a calendar
definition to translate a vector of numbers into a point in time.
}
