% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtime.R
\name{new_mixtime}
\alias{new_mixtime}
\title{Create a new mixtime}
\usage{
new_mixtime(x = numeric(), calendar = new_calendar())
}
\arguments{
\item{x}{The iterable value of the mixtime (for example, a value of 1 may indicate the first day).}

\item{calendar}{The calendar structure over which the mixtime iterates. This can be created using \code{new_calendar()}.}
}
\description{
A mixtime is a vector which describes a point in time. It uses a calendar
definition to translate a vector of numbers into a point in time.
}
