% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_unit.R
\name{new_time_unit}
\alias{new_time_unit}
\title{Create a new time unit}
\usage{
new_time_unit(x, class)
}
\arguments{
\item{x}{The number of units (must be an integer of length 1).}

\item{class}{The class of the unit. Recommended class format is "tu_*", for
example, "tu_month".}
}
\value{
A time unit.
}
\description{
\lifecycle{experimental}
}
\details{
A time unit describes the granularity of a calendar. The time unit class is
used to identify how different time units nest another.
}
