% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time_unit.R, R/time_unit.R
\name{time_unit_full}
\alias{time_unit_full}
\alias{time_unit_abbr}
\title{Time units as a string}
\usage{
time_unit_full(x, ...)

time_unit_abbr(x, ...)

time_unit_full(x, ...)

time_unit_abbr(x, ...)
}
\arguments{
\item{x}{A time unit object (e.g., \code{tu_month(1L)})}
}
\value{
A string representing the time unit

A string representing the time unit
}
\description{
These S7 generic functions provide the full and abbreviated names for time
units.

These S7 generic functions provide the full and abbreviated names for time
units. This is used internally for the display of continuous time, cyclical
time, and durations.
}
\examples{
time_unit_full(tu_year(1L))

time_unit_full(tu_year(1L))
time_unit_abbr(tu_year(1L))

}
