% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{tz_offset}
\alias{tz_offset}
\title{Get timezone offset}
\usage{
tz_offset(time)
}
\arguments{
\item{time}{A POSIXct datetime object or something coercible to POSIXct.
The timezone is extracted from this object.}
}
\value{
A numeric vector of offsets in seconds from UTC.
}
\description{
Returns the UTC offset for a given datetime in its specified timezone.
}
\examples{
tz_offset(Sys.time())
tz_offset(as.POSIXct("2024-06-15 12:00:00", tz = "America/New_York"))
}
