% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso8601.R
\name{yearweek}
\alias{yearweek}
\title{ISO 8601 year-week time representation}
\usage{
yearweek(.data, tz = "UTC")
}
\arguments{
\item{.data}{Another object to be coerced into ISO 8601 year-weeks.}

\item{...}{Arguments for methods.}
}
\description{
Create or coerce using \code{yearweek()}.
}
\details{
\lifecycle{experimental}
}
\examples{

yearweek(Sys.Date())
yearweek(0:52)

}
